ec25/requirements.txt
# File: requirements.txt
# Dependencies untuk EC25 MQTT Service

pyserial==3.5
paho-mqtt==1.6.1